Do the math operations at the end.
Wait until the end to do inverse than doing it at each step because the order of operations needs to be in the correct order.

If the bone chain is: A -> B -> C -> D

The problem is D's rest pose is A*B*C*Dlocal.

So you want to get Dlocal from D.