Fix parsing of UTF16LE strings
The code had been assuming everything was UTF8, but that's only the
case if encoding is 1. For an encoding of 0, this code first converts
the string to an array of char16_t, explicitly copying the data to
avoid any platform endianness issues, then parses it as UTF16. We
also always pass the length to avoid any potential issues of reading
past the buffer if a string isn't null terminated.